requirements.txt
einops==0.8.0
jury==2.3.1
numpy==1.26.4
pandas==2.2.3
pillow==10.2.0
rotary-embedding-torch==0.8.6
scikit-learn==1.6.0
torch==2.1.1
--extra-index-url https://download.pytorch.org/whl/cu121
torchvision==0.16.1
tqdm==4.66.5
transformers==4.48.0
ratescore
